+++ MoveMoney.API/Models/ComissionRange.cs
@@ -5,8 +5,8 @@ namespace MoveMoney.API.Models
         public int Id { get; set; }
         public Comission Comission { get; set; }
         public int ComissionId { get; set; }
-        public decimal MinAmount { get; set; }
-        public decimal MaxAmount { get; set; }
+        public double MinAmount { get; set; }
+        public double MaxAmount { get; set; }
         public double Percentage { get; set; }
     }
 }
